// Files/src/js/models/scenario-id.ts
export enum ScenarioId {
	DUNGEON_RUN = 2663,
	MONSTER_HUNT = 2706,
	RUMBLE_RUN = 2890,
	DALARAN_HEIST_CHAPTER_1 = 3005,
	DALARAN_HEIST_CHAPTER_2 = 3188,
	DALARAN_HEIST_CHAPTER_3 = 3189,
	DALARAN_HEIST_CHAPTER_4 = 3190,
	DALARAN_HEIST_CHAPTER_5 = 3191,
	DALARAN_HEIST_CHAPTER_1_HEROIC = 3328,
	DALARAN_HEIST_CHAPTER_2_HEROIC = 3329,
	DALARAN_HEIST_CHAPTER_3_HEROIC = 3330,
	DALARAN_HEIST_CHAPTER_4_HEROIC = 3331,
	DALARAN_HEIST_CHAPTER_5_HEROIC = 3332,
}